Moved new read serial function
